fixed user-metadata in commons package
